issue #123 : Convert downloads to use thread pool executors
This restricts the total number of threads so it doesn't balloon.

Under pressure, there is a known maximum theoretical load on the system, with random latency increases to ensure that the load is always under the theoretical limit.

For each download, there are two ExecutorService's which will handle the download, one of which is the overall scheduler, chosen typically based on category, although not necessarily if a higher capacity scheduler is free. Then for each of the facets, there is another ExecutorService that restricts the number of facet specific threads from ballooning.

Signed-off-by: Peter Ansell <p_ansell@yahoo.com>