issue #123 : Enforce waiting for initialisation to complete
Ensures that we don't run into cases where downloads are lost because they are submitted early.

The delayed initialisation to parallelise application startup procedures is useful overall, but we need to put in wait points before servicing requests, which this provides using CountDownLatch with a single latch call to trigger it.

Signed-off-by: Peter Ansell <p_ansell@yahoo.com>