Use loop again in ScatterplotController
The Map and Set are both currently equivalent due to the Exceptions that are thrown, guaranteeing that either all fields with a given name are "isStored" or an exception is thrown.

However, a future fix for this will use the Set rather than the Map, so simpler to keep the iteration over the set in this case.

Adding a break statement when a match occurs to avoid future issues where a name matches multiple index fields and one of them isn't stored.

Signed-off-by: Peter Ansell <p_ansell@yahoo.com>