issue #199 : Improve error messages and always generate citations.csv if enabled
The error messages are hiding the cases where the collectory is responding with "null". In order to debug where the underlying issue with the "null" is, we need to know which uids were requested, which requires a change to the logging message to include that information.

Also separating the parameter sanity checks up so that the parameter which was incorrect can be identified from a stacktrace, so we don't need to guess which parameter may have been the issue.

Signed-off-by: Peter Ansell <p_ansell@yahoo.com>