Add configurable delay after DOI minting before sending emails containing the DOI
Users have commented that they have noticed issues with the DOI being resolvable immediately when they receive the email, and I have replicated the failures myself by attempting to resolve the DOI immediately after ANDS has accepted it and minted a DOI for it.

The cause is likely a propagation delay between ANDS and DataCite.

Empirically determined the delay to be around 30 seconds, so setting the default delay to 1 minute (60000 ms) to avoid future issues.\

Concurrent downloads in other pools are unaffected by this delay period, and given the offline/asynchronous nature, users will also not be affected by it.

Signed-off-by: Peter Ansell <p_ansell@yahoo.com>